requirements.txt
apache-airflow-providers-docker
pandas
s3fs
pyarrow
sqlalchemy
psycopg2-binary
dbt-core
dbt-postgres